+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:50266/07eecaf8-121c-4fb1-a75e-e531ef2881a7?"}
+{"mongoUri":"mongodb://127.0.0.1:50613/e1a14a4d-1e3a-4f2e-872b-4008b787febf?"}
