+++ src/main/factories/usecases/account/authentication/db-authentication-factory.ts
@@ -3,7 +3,7 @@ import { BcryptAdapter } from '@/infra/criptografy/bcrypt-adapter/bcrypt-adapter
 import { JwtAdapter } from '@/infra/criptografy/jwt-adapter/jwt-adapter'
 import { Authentication } from '@/domain/usecases'
 import { DbAuthentication } from '@/data/usecases'
-import { AccountsRepository } from '@/infra/db/typeorm/repositories/accountsRepository'
+import { AccountsRepository } from '@/infra/db/typeorm/repositories/account/accountsRepository'
 
 export const makeDbAuthentication = (): Authentication => {
   const salt = 12
